working on creating modals
