+++ react-app/src/components/CreateProduct/index.css
@@ -18,4 +18,9 @@
 
 .creater-product-button{
     background-color: #222222;
+    color: white;
+    width: 100%;
+    padding: 10px;
+    border-radius: 20px;
+    margin-top: 20px;
 }
