+++ react-app/src/components/OpenModalButton/index.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { useModal } from '../../contex/Modal';
+import { useModal } from '../../context/Modal';
 
 function OpenModalMenuItem({
     modalComponent, // component to render inside the modal
@@ -19,5 +19,26 @@ function OpenModalMenuItem({
         <div className='modal-button-text' onClick={onClick}>{itemText}</div>
         );
 }
+
+export function OpenModalButton({
+    modalComponent, // component to render inside the modal
+    buttonText, // text of the button that opens the modal
+    onButtonClick, // optional: callback function that will be called once the button that opens the modal is clicked
+    onModalClose // optional: callback function that will be called once the modal is closed
+  }) {
+    const { setModalContent, setOnModalClose } = useModal();
+
+    const onClick = () => {
+      if (typeof onButtonClick === 'function') onButtonClick();
+      if (typeof onModalClose === 'function') setOnModalClose(onModalClose);
+      setModalContent(modalComponent);
+    };
+
+    return (
+      <button onClick={onClick}>{buttonText}</button>
+    );
+  }
+
+
 
 export default OpenModalMenuItem;
