+++ react-app/src/index.js
@@ -5,26 +5,20 @@ import { BrowserRouter } from 'react-router-dom';
 import './index.css';
 import App from './App';
 import configureStore from './store';
-import ModalProvider, {Modal} from './contex/Modal';
+
+import { ModalProvider,Modal} from './context/Modal';
 
 const store = configureStore();
 
-function Root() {
-  return (
+ReactDOM.render(
+  <React.StrictMode>
     <ModalProvider>
       <Provider store={store}>
         <BrowserRouter>
           <App />
-          <Modal />
         </BrowserRouter>
       </Provider>
     </ModalProvider>
-  );
-}
-
-ReactDOM.render(
-  <React.StrictMode>
-    <Root />
   </React.StrictMode>,
   document.getElementById('root')
 );
